+++ CHANGELOG.md
@@ -7,6 +7,36 @@ All notable changes to this project will be documented in this file.
 
 * Support for `UICollectionViewDelegate.collectionView(_:canPerformPrimaryActionForItemAt:)` and `UICollectionViewDelegate.collectionView(_:performPrimaryActionForItemAt:)` delegate methods on iOS 16 and tvOS 16.
 * Support for `UICollectionViewDelegate.collectionView(_:contextMenuConfigurationForItemsAt:point:)`, `UICollectionViewDelegate.collectionView(_:contextMenuConfiguration:highlightPreviewForItemAt:`) and `UICollectionViewDelegate.collectionView(_:contextMenuConfiguration:dismissalPreviewForItemAt:` methods on iOS 16.
+* Support for `UIHostingConfiguration` on iOS 16 / tvOS 16 / macCatalyst 16:
+
+```swift
+manager.registerHostingConfiguration(for: Post.self) { _, post, _ in
+    UIHostingConfiguration {
+        PostView(post: post)
+    }
+}
+```
+
+It's also possible to incorporate UIKit cell states by simply adding additional parameter to registration:
+
+```swift
+manager.registerHostingConfiguration(for: Post.self) { state, _, post, _ in
+    UIHostingConfiguration {
+        PostView(post: post, isSelected: state.isSelected)
+    }
+}
+```
+
+Additionally, it's possible to customize `UICollectionViewCell` being used to host SwiftUI view, for example for list cells:
+
+```swift
+manager.registerHostingConfiguration(for: Post.self, cell: UICollectionViewListCell.self) { _, post, _ in
+    UIHostingConfiguration {
+        PostView(post: post)
+    }
+}
+```
+
 * Support for events, wrapping `UICollectionViewDataSourcePrefetching` protocol. 
 
 ```swift